third_party/WebKit/LayoutTests/http/tests/security/window-named-proto-expected.txt

CONSOLE WARNING: Resource requests whose URLs contain raw newline characters are deprecated, and may be blocked in M60, around August 2017. Please remove newlines from places like element attribute values in order to continue loading those resources. See https://www.chromestatus.com/features/5735596811091968 for more details.
CONSOLE ERROR: line 2: Uncaught TypeError: Cannot read property 'innerHTML' of null
